refactor: 입장하기, 개설하기 페이지 스텝별 컴포넌트화
